+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Excluded phpunit.xml.dist from `git archive`
 * Use `spl_object_id` whenever it is possible
 * Do not add `...` to the end of text whenever text is not shortened
+* Added constants `Awesomite\VarDumper\LightVarDumper::DEFAULT_*`
 * Allow displaying array in single line whenever array contains only 1 element,
 element has index 0 and output is not too long e.g.
 ```